requirements.txt
# FastAPI and server
fastapi
uvicorn[standard]

# Database drivers
asyncpg
sqlalchemy
alembic

# Vector library
pgvector

# AI/ML libraries
openai
anthropic

# Data validation and settings
pydantic
pydantic-settings

# HTTP and file handling
httpx
python-multipart

# Authentication and security
python-jose[cryptography]
passlib[bcrypt]

# Environment and configuration
python-dotenv

# Caching
aioredis

# Utilities
python-dateutil

openai>=1.0.0
anthropic>=0.25.0
python-dotenv>=1.0.0